fix useFormFields test
