+++ packages/forms/babel.config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  ...require('../../babel.base.js'),
+};
