+++ packages/forms/package.json
@@ -20,11 +20,12 @@
   ],
   "scripts": {
     "build": "rm -rf dist && tsc -p ./tsconfig.build.json",
-    "typecheck": "tsc --noEmit"
-  },  
+    "typecheck": "tsc --noEmit",
+    "test": "jest"
+  },
   "dependencies": {
-    "@react-native-ama/internal": "*",
     "@react-native-ama/core": "*",
+    "@react-native-ama/internal": "*",
     "@react-native-ama/react-native": "*"
   },
   "peerDependencies": {
@@ -49,5 +50,33 @@
     "provenance": true
   },
   "author": "",
-  "license": "MIT"
+  "license": "MIT",
+  "jest": {
+    "preset": "jest-expo",
+    "modulePathIgnorePatterns": [
+      "<rootDir>/node_modules",
+      "<rootDir>/dist/"
+    ],
+    "collectCoverageFrom": [
+      "src/**/*.{ts,tsx}"
+    ],
+    "coverageThreshold": {
+      "global": {
+        "statements": 90,
+        "branches": 82,
+        "functions": 83,
+        "lines": 91
+      }
+    },
+    "transformIgnorePatterns": [
+      "node_modules/(?!(@react-native|react-native|react-native-reanimated/))"
+    ],
+    "transform": {
+      "\\.[jt]sx?$": "babel-jest"
+    },
+    "verbose": true
+  },
+  "devDependencies": {
+    "jest-expo": "^49.0.0"
+  }
 }
